refactor: centralize terminal input handling
